add body panel
